refactor(src): split monolithic main.rs into focused modules
Extract PresetFile, drain functions, and three UI panels from main.rs
into separate modules: src/presets.rs, src/ui/drain.rs,
src/ui/timeline.rs, src/ui/clip_browser.rs, and src/ui/monitor.rs.
src/ui/mod.rs serves as a pure re-export file. main.rs is reduced
to a 57-line thin coordinator.
